Tweaked Priority Hints implementation to work across more content types
In preparing to re-start the priority hints origin trial, this provides
control for tweaking the priority of most content types and adds some
more control to the existing types that were covered before.

Doc with the full matrix for resource priorities including hints:
https://docs.google.com/document/d/1bCDuq9H1ih9iNjgzyAL0gpwNFiEP4TZS-YLRp_RuMlc/edit?usp=sharing

The getResourcePriority used by tests was changed to get the initial
priority when the request is created instead of the priority when the
request is sent. This is because the priority changes as the main html
parser reaches script and stylesheets in the body (boosting priority)
and for visible images. The tests need to be able to test the
priority that was set when the request was created without the
side-effects of priorities changing as images become visible or the
main parser traversing the document.

The tests themselves were also expanded to cover more cases,
particularly late-body scripts and stylesheets which are treated
differently than those in the head and some of the existing tests
were fixed (not using a const in the for loop was causing them to
not use the expectations from the defined block and some were not
checking the correct urls).

Bug: 821464
Change-Id: I3077844d7ec10f3681f79383d0f9f05744fc0497
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3146030
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Patrick Meenan <pmeenan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#922137}
NOKEYCHECK=True
GitOrigin-RevId: 9a25598504d62ab5ec2a6a96a471012a5dabc6b1