Make LiteralDebugNameClient GarbageCollected
DisplayItemClient is inherited by LayoutObject, which was moved to Oilpan.
It is necessary to make it GarbageCollected because it’s dangerous to save a raw pointer to an on-heap object.
Every child class of DisplayItemClient should be GarbageCollected to make it possible.

Bug: 1248357
Change-Id: I6432b5a9aa90f97c0e32b8cf4a51023b9da6792c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3156310
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Yuki Yamada <yukiy@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Naho Tomiki <nahotomiki@google.com>
Cr-Commit-Position: refs/heads/main@{#922418}
NOKEYCHECK=True
GitOrigin-RevId: 16113fede11516ed85c18bba8bdfe7cb1a8c1145