Make inner class of PaintLayerScrollableArea GarbageCollected
DisplayItemClient is inherited by LayoutObject, which was moved to
Oilpan.
It is necessary to make it GarbageCollected because it’s dangerous to
save a raw pointer to an on-heap object.
Every child class of DisplayItemClient should be GarbageCollected to
make it possible.

Bug: 1248357
Change-Id: I7ae027d530a5d185d83558f39c84529b3e9d6984
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3159915
Reviewed-by: Yuki Yamada <yukiy@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Naho Tomiki <nahotomiki@google.com>
Cr-Commit-Position: refs/heads/main@{#922421}
NOKEYCHECK=True
GitOrigin-RevId: 4e2e3ae20b0961b6dc1903f6d0f63a02f172352e