content-visibility: Don't expose NearestLockedExclusiveAncestor.
This function is being used to check for layout blocking and paint
blocking and maybe other cases too. However, when we move to the model
that allows some of the phases to happen, then it will be confusing
to figure out whether this is the right function to use.

This patch changes all uses to an explicit LockedAncestorPreventing*
functions.

Also cleans up unused functions.

R=chrishtr@chromium.org

Change-Id: I6fc4d9c5397d071aced7578ffa2680b38778666d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3165358
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#922615}
NOKEYCHECK=True
GitOrigin-RevId: 4b13e65c914599d4be74a9f4db257fa1218b865b