Treat <header>/<footer> exposed as banner/contentinfo landmarks
An earlier code simplification in CL:2911373 caused a change in how
<header> and <footer> are exposed, resulting in a regression in our
NVDA support. The issue was that IsLandmark() does not contain
Role::kHeader or Role::kFooter, and therefore the xml-roles attribute
was not exposed for these.

According to HTML-AAM/CORE-AAM, a <header>/<footer> scoped to the
<body> should be receive banner/contentinfo landmarks.

Bug: None
Change-Id: I2978e364bc6296a8d918d74cd719043f239b9fb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3169191
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#922689}
NOKEYCHECK=True
GitOrigin-RevId: 8d16d942248c779bd2b841c6a3ce3b3364e3378c