[SelectMenu] Apply controller code to the author provided shadow DOM.
This change ensure that controller code is applied to the author
provided shadow DOM by having the mutation observer watch
changes for the shadow root.

Because of that we can't rely on the slotchange event to catch
when the default parts were replaced as we would have duplicate
notifications. In order to do that
HTMLSelectMenuElement::SelectMutationCallback should be extended to
handle the case when a node with slot attribute is inserted/removed
even if its part attribute isn't set.

Bug: 1121840
Change-Id: I20ec7a00b8a3a130740a4dc0354bba0563ebb0a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3166023
Reviewed-by: Dan Clark <daniec@microsoft.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#922716}
NOKEYCHECK=True
GitOrigin-RevId: d47dff1152bf5d91c0972366bca802edd5612962