[CFFU][#1] Ruggedize media_devices code
(CFFU is short for "Conditional Focus Follow-Ups")

Ruggedize the code in media_devices.h/cc to take into
account the possibility of a track being removed from
its stream before the microtask which closes the
focus-window fires.

The old code worked, because the MediaStream's ID
is what's associated with the source on the browser
process, but this code is clearer and more obviously
correct, making things more future-proof.

Bug: 1215480
Change-Id: I63e7bdb5466351f2fcf57c81d409a82b23ee5fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3165037
Commit-Queue: Elad Alon <eladalon@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#922926}
NOKEYCHECK=True
GitOrigin-RevId: a3de2db3a54ee4ff1f1dae151b934f54643b3e68