[DOM Storage + StorageKey] (2) Add StorageKey to IPC
This CL takes the first of two steps in IPC changes from the StorageKey
migration. The step taken here is the lower risk one as (for the moment)
StorageKey is a wrapper for Origin. The following step (adding
LocalFrameToken and verifying it) was the source of the errors last
time.

The prior attempt at this change (https://crrev.com/c/3122751). Took a
quick and dirty approach which required a revert. This code is on a
critical path, has a lot of quick fixes layered, and is under-tested.
Let's take a slower approach as follows:
(1) Refactor session and local storage checks
(2) Add StorageKey to IPC
(3) Add LocalFrameToken to IPC and verify

Bug: 1212808
Change-Id: I03242c9739f101206c1218239a8ee6e1fbb82d1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3166025
Commit-Queue: Ari Chivukula <arichiv@chromium.org>
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923152}
NOKEYCHECK=True
GitOrigin-RevId: b54d90420260245961929f43e8fb19f1a2e62e6a