[editing] Unify justifyright?4001-last expectations
The platform-specific expectations for linux and win10.1909 were the
same, and win produces the same result, but was marked as failing due to
an of date expectation.

This patch adds a non-platform-specific expectation for all of these,
and only keeps a different mac expectation.

Bug: 1243128
Change-Id: Ib06014dc0088ba97b4b7bd98cef3a1579a289af8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3162121
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/main@{#923163}
NOKEYCHECK=True
GitOrigin-RevId: 221801cffd16404d67b815e3db5cde3ada0d984b