BackForwardCache: Reland: Disable when MediaSession's service is used
This is reland of crrev.com/c/3121727 with disabling the test on the
ASAN bot (crbug.com/1243159). We are addressing on this issue soon
later.

After this change, back-forward cache is blocked when MediaSession's
service exists, i.e., a MediaSession action handler is registered and a
audio or media plays. It is risky to put a page with a MediaSession
service into the back-forward cache, since the page's state can affect
a media control UI. This change adds this blocking behind a flag
(supported_features) for future experiments.

This change adds feature flags. We think that changing the playback
state without a MediaSession service should not be harmful for back-
forward cache. By adding the flags, we will be able to do experiments.

After this change, we have two new features:

 * BackForwardCacheMediaSessionPlaybackStateChange
 * BackForwardCacheMediaSessionService

These are independent and can take these pairs:

  * (false, false): MediaSession's playback state change doesn't allow
                    BFcache (default)
  * (true, false):  MediaSession's playback state change w/o a
                    MediaSession service allows BFcache
  * (true, true):   MediaSession's playback state change w and w/o a
                    MediaSession service allows BFcache

In theory there is a pair (false, true) but this doesn't make sense, as
a MediaSession service does nothing if the playback state is not
changed.

Bug: 1243599
Change-Id: I01f1498aaf3796ade1c5a4520834eab434229902
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3151103
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923248}
NOKEYCHECK=True
GitOrigin-RevId: 568bc4acd1bcfbb506741c66f09ea9c9b9cda535