Don't include BR clearance in the line's height.
Including clearance in the line height may result in tall monolithic
content, which is problematic for block fragmentation, especially for
column balancing.

Instead, store trailing clearance in the layout result, so that the
block layout algorithm can add it to the layout position for subsequent
content.

We already use a similar trick for ruby annotations, and some care is
required so that we don't move back above the clearance offset when
canceling ruby annotations overflow (which happens in some cases). There
was no test coverage for this, so added three tests for that (they pass
both and without this CL).

Also added a multicol test that demonstrates what this fixes in column
balancing.

Bug: 829028
Change-Id: I84e58326e9ada78e3f29ad01d8043a173626aea3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3168645
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923502}
NOKEYCHECK=True
GitOrigin-RevId: 1c62e98ed9f24f70a188fd51d7c6c9f3aba47967