[mq-4] Make MediaQueryExpValue a type + union
The current implementation of MediaQueryExpValue::Equals seems
wrong. It assumes that if the current value type is X, then the other
value type must also be X, which means that two values of different
types can be equal. Also, it does not check |unit|, which means e.g.
10px and 10em are equal.

This CL fixes MediaQueryExpValue::Equals, renames it to operator==,
and while at it converts MediaQueryExpValue to a union.

Bug: 1034465
Change-Id: I6d4f785d2cd5245365a0c2b859395479a1483965
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3168877
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923580}
NOKEYCHECK=True
GitOrigin-RevId: ebd9ca849c5365a80b216c08bc41cadcdc87f1fa