Extract WebBundleLoader from LinkWebBundle
This is separated from https://crrev.com/c/3128843. We can land the
parts of the CL which don't have to be blocked by the on-going spec
discussion.

This CL factors out WebBundleLoader so that it is used from other
than LinkWebBundle.

Separating it now also reduces the maintenance burden of
https://crrev.com/c/3128843 since rebasing the CL is likely to miss
a change of WebBundleLoader in link_web_bundle.cc on ToT.

Bug: 1245166
Change-Id: Ic10006b8bd60664aac9630dd4238ed5e7a656e74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3172057
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923661}
NOKEYCHECK=True
GitOrigin-RevId: 9a88bfa543d7a681c0f4a0a9ce8708effb3179bf