BackForwardCache: Forbid playing a media
Now there is an issue that a media controller UI is kept shown even
after a navigation and the old page with a playing media goes to back-
forward cache. In general, a media controller UI is shown when a media
like a video or an audio plays. A Media Session API usage doesn't matter
here.

As a temporary solution, let's forbid all the media play so that a page
with a playing media is not put into the cache. After this is landed,
let's fix the media controller UI to show and hide it in appropriate
timings in order to allow playing a media by default in the future.

Note that this is orthogonal with Media Session change
(crrev.com/c/3151103), and an actual media play matters in this CL. For
example, Media Session's playback state can be updated without an actual
media play.

Bug: 1246240
Change-Id: I81d24a608532ed918fa6649b96d744a755520fee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3151048
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923712}
NOKEYCHECK=True
GitOrigin-RevId: 980e2fd442ccb474a70c605083e43f708404c3b7