BackForwardCache: Extend the UMA Event.PageShow.Persisted to be recorded in browser
We found that these metrics values don't agree:

 * BackForwardCache.AllSites.HistoryNavigationOutcome=Restored (browser)
 * Event.PageShow.Persisted=True (renderer)

Both should be recorded when a page is restored from back-forward cache.
To investigate this issue more, let's add more metrics. This CL adds a
new enum EventPageShowPersisted and have the existing UMA
Event.PageShow.Persisted to record them instead of a boolean.

EventPageShowPersisted consists of:

 * kNoInRenderer = 0 (formerly false)
 * kYesInRenderer = 1 (formerly true)
 * kYesInBrowser = 2

Thus, the backward compatibility should be kept. kYesInBrowser is
recorded when browser triggers a pageshow event with persisted flag.

We don't have kNoInBrowser. In browser, Event.PageShow.Persisted is
recorded only when the pageshow event is persisted, but we will be able
to compare the numbers of kYesInRenderer and kYesInBrowser.

We will add more metrics later when necessary.

Bug: 1234634
Change-Id: Ifcd6cbf4ad4e91b80aa8730a97b67f7f8ac4e2b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3159920
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Yuzu Saijo <yuzus@chromium.org>
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923740}
NOKEYCHECK=True
GitOrigin-RevId: b8edd06c1520da2da1fd36a023bdeef4abfd07f6