Reland "[@layer] Fix cascade ordering of inline style and layered style"
This is a reland of crrev.com/c/3167095 with a bug fix that,
presentational hints should be added to match result without setting the
is_inline_style flag.

Test case presentational-hints-cascade.html somehow passed in the
original patch with the bug due to an experimental feature
CustomElementDefaultStyle. This patch adds it to the virtual/stable test
suite to make sure it fails without the bug fix.

Original description:

This patch adds a new flag to MatchResult and CascadePriority to
indicate if a declaration is from inline style or other style sheets, so
that we can compare inline style and other layered style correctly.

Without this flag, currently we are treating inline style as regular
declarations in the implicit outer layer, and therefore mistakenly
override it with layered style rules in other style sheets.

Bug: 1095765
Change-Id: Ibeb0ed2d7e571c136bb4a636b99a766fc9b6e4b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3174310
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924101}
NOKEYCHECK=True
GitOrigin-RevId: f2423df18f36b393444fddd112a0c96629b8e85d