[mq-4] Add some tests to MediaConditionParserTest.Basic
MediaQueryParser::ParseMediaCondition is a special mode of the
MediaQueryParser which is supposed to parse only a single
<media-condition> rather than a full <media-query-list>. It is used
to implement the 'sizes' attribute [1].

The current parser seems to get confused sometimes, allowing values
with a comma, but I'm adding those cases anyway as a baseline for
how the current parser works.

This is a preparation for readying the parser for Media Quries Level 4
syntax.

[1] https://html.spec.whatwg.org/multipage/images.html#sizes-attribute

Bug: 962417
Change-Id: Id4fed7c2f2ab6f9987ffa4da308ccc0044c1db50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3168653
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924207}
NOKEYCHECK=True
GitOrigin-RevId: f773d1ce29dbb878a15e705a0610fd12119573f8