Mark for whitespace reattachment based on layout object parent
The existing code marked based on light tree parent, which did not work
correctly for slotted elements. Instead keep track of the last layout
object parent which got a child removed during the detach part of a DOM
or flat tree removal.

The marking is only done for the closest LayoutObject parent of the
detached subtree. If the marking was done for every LayoutObject removal
in the subtree, the style recalc root would not be updated correctly and
we would have a less optimal recalc root as we try to update the recalc
root after the removal is complete.

Fixes two correctness issues and prepares for using the same framework
for doing the same for list item updates.

Bug: 1244080, 1247079, 1239137
Change-Id: Iacfe010c50f3f33a5e45b3ea996a3f42cb465f84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3172622
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924258}
NOKEYCHECK=True
GitOrigin-RevId: d59452974e3fd468597e972b68ec86c05f970f64