[CSP] Fix WPT about worker violation event
The WPT
external/wpt/content-security-policy/connect-src/worker-from-guid.sub.html
was expecting a securitypolicyviolation event for a Content Security
Policy violation in a worker to be raised in the owner document. This
is wrong: the securitypolicyviolation event should be triggered in the
worker itself.

Bug: 694525
Change-Id: I4285e6eee73e4629a6662a05fe2eb067782f6c8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3178905
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924287}
NOKEYCHECK=True
GitOrigin-RevId: 1bf1264aa947a87d157aa72a34c2c0d7a4e20848