[selectmenu] Remove re-entrant layout for selectmenu popup positioning
HTMLPopupElement::AdjustPopupPositionForSelectMenu uses the
LocalDOMWindow innerWidth/innerHeight getters, which can trigger a
relayout of the current document if viewport is enabled. Since we're
already doing layout at this point, this triggers a CHECK.

Update HTMLPopupElement::AdjustPopupPositionForSelectMenu to get the
size from the LocalFrameView directly such that the recursive
layout recalc is avoided.

The test included here is a web_test and not a WPT because the extra
same-document re-layout is only done when viewport is enabled, and
the desktop browser ignores <meta name="viewport">. So the test needs
to use  internals.settings to enable viewport instead.

Bug: 1252600
Change-Id: I953822deb9349ff1fca822ced4e032c7a6a47c41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3177072
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924798}
NOKEYCHECK=True
GitOrigin-RevId: 43c057d8e0fc346952e7bf8a69e996dc248f9079