Prevent CHECK on inspector paint layers lookup
As a result of crbug.com/1058527, certain combinations of Inspector
platform requests can yield a renderer crash. This occurs with certain
timings when the DOM domain requests paint layer information.
Simply updating `UpdateLifecycleToLayoutClean` to
`UpdateLifecycleToCompositingClean` didn't resolve the issue reliably.

Although it is likely preferable to address the issue of the inputs
being out-of-date, this change is a narrow fix to prevent crashing
while preserving the previous semantics, in which the functional
changes would simply return a default value (in this case, 0).

Bug: 1248251
Change-Id: Ia223f0951b1813f48351c7150e0261f6f20abdc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3154383
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#924821}
NOKEYCHECK=True
GitOrigin-RevId: 3249c5175069c7e4d016c1aa3796cdb78ac6e764