Remove "resize" as a <popup> light dismiss trigger
This ended up being a confusing part of the implementation of
<popup>: it was easy to accidentally resize the popup, causing
it to light dismiss. For example, adding a :hover style that
adds a border to the popup or an element might cause a light
dismiss. And the rationale for this being a trigger in the
first place was an implementation concern - avoiding
expensive relayout in the case that the popup is anchor-
positioned to something on the page.

Bug: 1168738
Change-Id: Ic5a3e616941ace8b47b4abc02dbf786aa78521c4
Fixed: 1252176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3180152
Commit-Queue: Mason Freed <masonf@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#924911}
NOKEYCHECK=True
GitOrigin-RevId: 553ce90c96e04944e5515404fc1ed17b049b5f6c