EditContext: implement attachedElements()
This CL implements EditContext.attachedElements() which returns elements
that are associated with the EditContext.

Related w3c issue: https://github.com/w3c/editing/issues/300

The CL also makes change to Element::setEditContext so only elements
in active document can be attached to EditContext. And when an element
is removed from tree, it'll also be disconnected from associated
EditContext (if it has any)

4 tests are added in edit-context.html:
'.attachedElements() should return associated elements',
'EditContext can only be associated with an element
that is in an active document.',
'If an element is removed from tree, it will disconnect any
associated EditContext.'
'If an element's ancestor is removed from tree, the element
will disconnect its associated EditContext.'

Minor changes:
1. rename blur/focus to Blur/Focus per feedback in
https://chromium-review.googlesource.com/c/chromium/src/+/2794801
2. fix typo in the test 'Testing Element.editContext'

Note when the focus is switched between two elements that share same
EditContext, the EditContext shouldn't be blurred and re-focused.
This will be addressed in a separate CL.

Bug: 1222639
Change-Id: I1823691f6fad9bf07a69a2c61c91abd9e91259e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3180103
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Alex Keng <shihken@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#925150}
NOKEYCHECK=True
GitOrigin-RevId: 3b97a131c1d06e582805adf3d2911870771915cc