[wasm] Add test for cross-origin module sharing
This CL adds two more tests for cross-origin sharing of WebAssembly
modules. The first test tries to send a module to an iframe in a
different agent cluster. The second test tries to send a module to an
iframe in the same agent cluster. In both cases the
postMessage should fail.

R=jbroman@chromium.org

Bug: chromium:1225641
Change-Id: Id1046cab521f65b68ab15bf55faee54cda70160e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3169714
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#925209}
NOKEYCHECK=True
GitOrigin-RevId: c786cce11b70d4be492941761622dd0e2f023db4