Clean up and refactor CopyConfigsAndCleanup
Change the parameter names from:
  def CopyConfigsAndCleanup(config_dir, dest_dir):
to:
  def CopyConfigsAndCleanup(src_dir, dest_dir):

Since the caller calls this function as:
  CopyConfigsAndCleanup(temp_dir, config_dir).
it is confusing that 'config_dir' is both the name of the first formal
parameter and the name of the second actual parameter.

Also move the shutil.rmtree() call to the caller so that the caller
creates and removes temp_dir. As a result, the function is renamed
"CopyConfigs".

Change-Id: Ic3168ce163f52ab735efe801c4077af2e10bfde3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3187854
Reviewed-by: Will Cassella <cassew@google.com>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Wan-Teh Chang <wtc@google.com>
Cr-Commit-Position: refs/heads/main@{#925382}
NOKEYCHECK=True
GitOrigin-RevId: 7972dad948142c13cec56e41ddcba221347b7b47