[ozone/wayland] Implemented support for xdg-foreign-v2.
This is the second part of [1] that is now split.  This CL adds support
for the new version of xdg-foreign protocol.

Chromium currently uses the protocol version 1, which is already on its
way to be replaced with the newer version 2.  KDE Plasma uses the v2
protocol already, and GNOME has is pending [2].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3177662
[2] https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1043

Bug: 1126817
Change-Id: Ia7801522a081ee5768a95743ada8e705c44f8330
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3178916
Reviewed-by: Mitsuru Oshima (exo only) <oshima@chromium.org>
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/main@{#925635}
NOKEYCHECK=True
GitOrigin-RevId: 222d6bb217d2657a4b9b85b069096b6eb38c7f5d