Highlight API: Add test to check for repaint after CSS rule change
Add a test validating that a Highlight range is repainted if the
corresponding ::highlight CSS rule changes. Currently this works
correctly because changing the CSS causes ValidateHighlightMarkers
to remove and re-add all ranges to the DocumentMarkerController,
triggering a repaint. But, an optimization to ValidateHighlightMarkers
to eliminate this extra work could cause the repaint to be skipped.

So, adding a test to ensure that we don't regress this scenario.

Bug: 1251193
Change-Id: I12936364eb057db6c684df4b45703e5bd9cb5cb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3182785
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#925798}
NOKEYCHECK=True
GitOrigin-RevId: 1ab185827007faf274a61859ae21af71cc992e7e