[WebCodecs] Implement readback for VideoFrame::copyTo().
This CL reworks the readback implementation in video_util.cc to expose
reading a plane to a pre-allocated buffer, which copyTo() then uses to
copy to an output ArrayBuffer.

This change also adds support for more formats, such as I420, to the
readback implementation. The set of formats supported for copyTo() is
now the same as the set of formats supported for encoding.

Fixed: 1225036
Change-Id: Icba317337409634c54ee92d9a49e061c7b0037b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3183025
Commit-Queue: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#925965}
NOKEYCHECK=True
GitOrigin-RevId: e2c8b6114cdb84e26cd44c82d095eb6976795697