Re-enable and update DevTools long cookie warning test
We will re-enable the DevTools test that ensures a console
warning is generated when a cookie exceeding the max size
is detected (and keep that in place until we can replace
with a DevTools Issue).

This CL should pass the web tests once [1] lands.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3157774

Bug: 1252463
Change-Id: I6ef57c39085ac938cc78decb52232257a244a9d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3189493
Commit-Queue: Andrew Williams <awillia@google.com>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926085}
NOKEYCHECK=True
GitOrigin-RevId: eceb3cfeebe12bc4a61c8b1ab76a95cc31ade2f4