Add Isolation Mode to DevTools inspector overlay
This adds a new type of persistent overlay: isolation mode, to inspector
overlays. When it is turned on for an element, the overlay will:
1. find the given element's current width, height, and content box;
2. and use these information to position three resizers for width,
height, and both directions (similar to DevTools' device preview mode);
3. send these information and the positions of the three resizers
to the overlay frontend to draw.

When the overlay frontend draws the resizers, it will watch mouse events
on the resizers and send resizing messages to the overlay backend, which
will decode the message and adjust the element's width and/or height
accordingly.

There will be some follow-ups to solve some niche cases, e.g. when the
element is rotated, etc. These niche cases will be collected and we need
more UX work to determine what to do in these cases.

Design doc: https://goo.gle/devtools-container-queries

frontend CL: https://crrev.com/c/3175289

screencast: https://imgur.com/a/61Tvewc

Bug: chromium:1146422
Change-Id: I1b072d3ba255a37fac05b0f71e5d59529c566d2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3173212
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926181}
NOKEYCHECK=True
GitOrigin-RevId: ab24109ebff41d668fa343d90fc5e2b7f160f20e