Add ability to query all fenced frame elements in a document.
Much like the DocumentPortals object create a DocumentFencedFrames
object that is a fast cache of all the fenced frames in a document.
This will allow layout for remote frames to work correctly in the
future. For layout to work correctly we need to be able quickly
iterate the RemoteFrames in a document, for normal OOPIF this
is done via iterating the FrameTree, but for Portals and Fenced Frames
we have side caches so we don't need to do an entire DOM walk.

Bug: 1123606
Change-Id: I5f0cdc32f53c7c8698d5d21738c3f37f721a5ce1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3189773
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926272}
NOKEYCHECK=True
GitOrigin-RevId: e143f74cb80a780140fc06eef822a752b227c81f