Put the LCP invisible image optimization behind a flag
LCP is not currently calculated for images that are initially invisible,
for optimization purposes. This results in correctness issues, where the
LCP image is not being reported.
This CL puts that optimization behind a flag, so that we can experiment
with turning it off.

Bug: 1249622
Change-Id: I2d2bcb03747f874acab7d45c77a5d25d6d319bee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3178913
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926365}
NOKEYCHECK=True
GitOrigin-RevId: 91e1319b0028448170bc4948828b3fdfa6171055