Block external protocol handler with sandbox.
Developers are surprised that sandboxed iframe can navigate the user
toward an external protocol.
The navigation in the sandbox is not blocked by sandbox normally, only
top-level navigation and opening popups are blocked. So not blocking
iframe navigations in general is working as intended. However sandboxed
iframe navigation toward external protocol handler are navigating the
user toward a different application. Somehow, this can be seen as a new
popup or a new top-level navigation.

This patch gates access to external protocol behind one of:
- allow-popup
- allow-top-level-navigation
- allow-top-level-navigation-by-user-activation + UserGesture.

This is not enforced by content/. So this is not a web feature.
This is implemented by chrome/ only.

The reason is that Android Webview may do not want to apply these
restrictions. Android Webview do not support external protocol by
itself, but it let the Android app to listen and intercept the request.
We decided to let the Android app full power here, to avoid breaking
the existing users.

Spec:
https://github.com/whatwg/html/pull/7124

Mozilla position:
https://github.com/mozilla/standards-positions/issues/581

Webkit position:
https://lists.webkit.org/pipermail/webkit-dev/2021-September/031988.html

Chrome feature:
https://chromestatus.com/guide/edit/5680742077038592

Change-Id: I0c3f130d084e7a23555178f3a172efb3607e87b6
Bug: 1148777, 1253379
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3162979
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926684}
NOKEYCHECK=True
GitOrigin-RevId: ad1f7dbbaa99230cf8637fc035461466a3fc0237