Make GraphicsLayer and GraphicsLayerClient GarbageCollected
DisplayItemClient is inherited by LayoutObject, which was moved to
Oilpan.
It is necessary to make it GarbageCollected because it’s dangerous to
save a raw pointer to an on-heap object.
Every child class of DisplayItemClient should be GarbageCollected to
make it possible.
Among them, GraphicsLayer and FrameOverlay becomes GarbageCollected by
this change.

Bug: 1248357
Change-Id: I5603f5e8f956390798d69f2fa1ff34328b8b3b96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3162940
Commit-Queue: Naho Tomiki <nahotomiki@google.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Yuki Yamada <yukiy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926694}
NOKEYCHECK=True
GitOrigin-RevId: b8a060648ed4d5dc00e0f6ee7d76165c24c5d96a