Use IsOffscreenCanvasRegistered to check if an OffscreenCanvas exists
HTMLCanvasElementModule::transferControlToOffscreen has been using the
existence of SurfaceLayerBridge to stand in for the existence of an
offscreen canvas. This is an incorrect assumption because frameless
canvases do not have surface layer bridges. Instead we should use the
more explicit IsOffscreenCanvasRegistered. The prevents the crash in
the linked bug.

Bug: 1254125
Change-Id: I37e9971eeb3447820445a81c2d4ca425ceac6515
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3193314
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926859}
NOKEYCHECK=True
GitOrigin-RevId: 2409d15eb012145f9a018a0e69843cf74d9d532c