Free video frame buffers when MediaStream renderer is suspended
Previously GpuMemoryBufferVideoFramePool was kept even when the stream
is not being rendered on the screen. As result it was still consuming
significant amount of memory for these buffers. Updated WebMediaPlayerMS
to free the pool when the renderer is suspended.

Bug: 1180323
Change-Id: I9a543d22e66e6258e6b981bf20aeba23e0691068
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3191071
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927045}
NOKEYCHECK=True
GitOrigin-RevId: d4b5ce0e6db1cdef52f89be05a8bcec16c4fe3a7