[block-in-inline] Mark ruby-justification-001.html as failure
This patch marks:
  external/wpt/css/css-ruby/ruby-justification-001.html
as failure.

This test is to test inlinification of ruby text using `<rtc>`
without `<rt>`. For browsers not supporting `<rtc>` yet, this
means we have a block-in-inline in the ruby base.

The current NG passes this test accidentally, failing to
justify the last line before block-in-inline. The
LayoutNGBlockInInline behavior is more correct for browsers
that do not support `<rtc>` yet.

Bug: 716930, 27659
Change-Id: I39edac3eb6429df0459edd2fdba456766edb3714
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3201939
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927612}
NOKEYCHECK=True
GitOrigin-RevId: a7807c4be8605a7d9f7b543c66437f617fe48dfe