Introduce WebTransport.Close and WebTransportClient.OnClosed
Introduce network.mojom.WebTransport.Close and
network.mojom.WebTransportClient.OnClosed. They are corresponding to
https://w3c.github.io/webtransport/#session-terminate and
https://w3c.github.io/webtransport/#session-terminated.

We have WPTs that tests this change in
/wpt/webtransport/close.https.any.js. This is skipped as the test server
infrastructure is not ready. Also, the underlying implementation in net/
is not ready so the tests fail even in my local environment.

Bug: 1011392, 1242902, 1242906
Change-Id: I460f1a18815dbeee3720a637d9f496a24aed36d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3197410
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927651}
NOKEYCHECK=True
GitOrigin-RevId: 17ea049b5c51554a2e78c6c4d90ce72115383817