[@container] Allow pseudo elements to query originator
Don't evaluate container queries for pseudo element rules when matching
rules for the originating element. If it turns out the originating
element is the container, we will use an incorrect evaluator. Also, the
flag for HasPseudoElementStyle() may not be set and the pseudo elements
never generated.

Instead assume the query evaluates to true and have false positives for
HasPseudoElementStyle().

Bug: 1217976
Change-Id: I6a80137564a45462c61d14e37f4e3e977ab40832
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198258
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927661}
NOKEYCHECK=True
GitOrigin-RevId: f5d0cef0606901f898dce424a990804e9bcdcf06