[loader] Add weakptr wrapper for CodeCacheHost
Add a mojo::Remote-owning wrapper for CodeCacheHost which allows
creating a WeakPtr to itself.

This WeakPtr can then be held by CodeCacheLoader, so that it can safely
access the CodeCacheHost with a way of handling the case when the
mojo::Remote is reset (the WeakPtr will be reset).

A side effect is that the interface passed to the code cache functions
is less mockable, so we have to introduce a more proper mojo message
pipe and corresponding task environment to some of the tests.

Bug: chromium:1213818, chromium:1245526
Change-Id: Iff02bcb3cd854c8f63de7a43e75c1ba52b07a468
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3162165
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928580}
NOKEYCHECK=True
GitOrigin-RevId: e2f98095874c96ca852ea92ccbd0f6c7acb4e66d