Re-enable fractional line-height
Turn on fractional line-height for layout measuring and also fix
and issue where we were flooring via returning an int. This other
codepath is used by getComputedStyle but also for measuring things like
the empty line for textarea, whereas the placeholder text would get the
correct fractional value. This mismatch was causing extra scrollbars
to appear. The caret-height.html test change was also made as the
previous test was incidentally (font-size: 10 is not valid, resulting in
the default font-size of 13.333px) hitting a case where the y position
of the caret was starting on a subpixel boundary when line-height was
set, and pixel snapping caused a difference in heights for the caret.

Change-Id: I6a0a295b25dcd75cf29aa89d7a885d8cbaf394b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3199976
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#928885}
NOKEYCHECK=True
GitOrigin-RevId: 878d6b63935a792c8601ec401121dec4730ee4a2