Revert 'Make calc() function can be used in any place that only accepts integer.'
This patch is a preparation step to support the behavior above per spec.
The original patch was landed as a first step to experiment,
but it was enclosed by the runtime flag "CSSCalcAsInt" as Anders suggested,
because the patch's behavior was not spec-compliant
and he concerned about if that patch's behavior is shipped
and sites start to relying on the incorrect CSSOM values,
it would be annoying to fix later.

So to make things clearer and to support spec-compliant behavior with separate CL,
I revert the change, https://chromium-review.googlesource.com/c/chromium/src/+/1478852.

Bug: 931216
Change-Id: I76d2eaa52dd0af44260b2034856437d9de43b147
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3208467
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Joonghun Park <pjh0718@gmail.com>
Cr-Commit-Position: refs/heads/main@{#928918}
NOKEYCHECK=True
GitOrigin-RevId: 69f817992e7c6c6338184d9cf538850f61c7cfc6